[HOTSPOT-74] refactor: DateNavigation 파일명 오타 수정
