+++ apps/web/src/features/report/index.ts
@@ -1,3 +1,3 @@
-export { DateNavigation } from './ui/DataNavigation';
+export { DateNavigation } from './ui/DateNavigation';
 export { ReportTypeSelector } from './ui/ReportTypeSelector';
 export { UserSelector } from './ui/UserSelector';
